refactor: move proxy record to new file
