+++ packages/mockyeah/app/proxyRoute.js
@@ -1,9 +1,8 @@
 const request = require('request');
 const isAbsoluteUrl = require('is-absolute-url');
 const { isEmpty } = require('lodash');
-const { decodeProtocolAndPort, handleContentType } = require('./lib/helpers');
-
-const now = () => new Date().getTime();
+const { decodeProtocolAndPort } = require('./lib/helpers');
+const proxyRecord = require('./lib/proxyRecord');
 
 const openingSlashRegex = /^\//;
 const leadUrlEncodedProtocolRegex = /^(https?)%3A%2F%2F/;
@@ -45,7 +44,6 @@ const makeRequestOptions = req => {
 
 const proxyRoute = (req, res, next) => {
   const { app } = req;
-  const { recordMeta } = app.locals;
 
   if (!app.locals.proxying) {
     next();
@@ -59,7 +57,7 @@ const proxyRoute = (req, res, next) => {
     return;
   }
 
-  const startTime = now();
+  const startTime = new Date().getTime();
 
   app.log(['proxy', req.method], reqUrl);
 
@@ -71,62 +69,14 @@ const proxyRoute = (req, res, next) => {
 
     if (!app.locals.recording) return;
 
-    const { options: { headers: optionsHeaders, only, useHeaders, useLatency } = {} } = recordMeta;
-
-    if (only && !only(reqUrl)) return;
-
-    const { method, body: reqBody } = req;
-
-    const { statusCode: status, _headers: __headers } = res;
-
-    const latency = now() - startTime;
-
-    let match = {
-      url: reqUrl
-    };
-
-    if (method && method.toLowerCase() !== 'get') {
-      match.method = method.toLowerCase();
-    }
-
-    if (reqBody && !isEmpty(reqBody)) {
-      match.body = reqBody;
-    }
-
-    if (optionsHeaders && Object.keys(optionsHeaders).length > 0) {
-      match.headers = Object.assign({}, optionsHeaders);
-    }
-
-    // If the match has only `url`, we can just serialize that as string.
-    if (Object.keys(match).length === 1) {
-      match = match.url;
-    }
-
-    const headers = Object.assign({}, __headers);
-
-    // Don't forward the suite header onto the proxied service.
-    delete headers['x-mockyeah-suite'];
-
-    // Don't record the `transfer-encoding` header since `chunked` value can cause `ParseError`s with `request`.
-    delete headers['transfer-encoding'];
-
-    let responseOptions = {};
-
-    if (status !== 200) {
-      responseOptions.status = status;
-    }
-
-    responseOptions = Object.assign(responseOptions, handleContentType(_body, headers));
-
-    if (useHeaders) {
-      responseOptions.headers = headers;
-    }
-
-    if (useLatency) {
-      responseOptions.latency = latency;
-    }
-
-    recordMeta.set.push([match, responseOptions]);
+    proxyRecord({
+      app,
+      req,
+      res,
+      reqUrl,
+      startTime,
+      body: _body
+    });
   })
     .on('response', _res => {
       if (app.config.responseHeaders) {
